requirements.txt
django==2.2
psycopg2==2.8
phonenumbers==8.10
django-phonenumber-field==3.0